pyproject.toml
[project]
name = "mdit-py-emoji"
version = "0.1.0"
description = "Emoji plugin for markdown-it-py."
authors = [
    {name = "BlueGlassBlock", email = "blueglassblock@outlook.com"},
]
dependencies = [
    "markdown-it-py<3.0.0,>=1.0.0",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.8.0",
    "isort>=5.10.1",
    "pytest>=7.1.3",
    "pytest-cov>=3.0.0",
]

[tool.pytest.ini_options]
python_files = "test.py"
addopts = "--cov=mdit_py_emoji --cov-report xml"
